Square 'Create New' button as first item in asset list
