+++ theme/asset-editor.less
@@ -74,7 +74,8 @@
     align-content: flex-start;
 }
 
-.asset-editor-card {
+.asset-editor-card,
+.create-new {
     width: 8rem;
     height: 8rem;
     margin: 1rem 0 0 1rem;
@@ -135,8 +136,19 @@
 .create-new {
     color: @white;
     background-color: @green;
-    margin-left: auto;
     font-weight: 700;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    border-width: 0;
+
+    i.icon {
+        margin: 0;
+    }
+}
+.create-new:hover {
+    background-color: darken(@green, 10%);
 }
 
 .delete-asset {
