+++ webapp/src/components/assetEditor/assetCardList.tsx
@@ -11,6 +11,7 @@ export class AssetCardList extends React.Component<AssetCardListProps> {
     render() {
         const { assets } = this.props;
         return <div>
+            {this.props.children}
             {assets.map(asset => <AssetCard asset={asset} key={asset.id} />)}
         </div>
     }
